// src/core/enums/notification.ts
export enum NotificationType {
  LIKE_MY_POST = 'LIKE_MY_POST',
  LIKE_MY_COMMENT = 'LIKE_MY_COMMENT',
  NEW_POST_FROM_FRIEND = 'NEW_POST_FROM_FRIEND',
  USER_COMMENTED_MY_POST = 'USER_COMMENTED_MY_POST',
  NEW_REQUEST_FRIEND = 'NEW_REQUEST_FRIEND',
  USER_ACCEPTED_REQUEST_FRIEND = 'USER_ACCEPTED_REQUEST_FRIEND',
}